Expect bytes in create_ and verify_signature; do not serialize
The caller knows what their data is and how to serialize it (including
whether or not to canonicalize it in some fashion); ssl.keys is too
low level to be doing this.  (I talked this over with Lukas and it
serves his interests for in-toto as well.)

Updated testing.  Did not touch code style, even though I'd love to....

Signed-off-by: Sebastien Awwad <sebastien.awwad@gmail.com>